Extract pagination control
