+++ src/ServicePulse.Host/vue/src/components/DataView.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { computed, onMounted, onUnmounted, ref } from "vue";
 import { useFetchFromServiceControl } from "../composables/serviceServiceControlUrls";
-
+import Pagination from "../components/Pagination.vue";
 const props = defineProps({
   apiUrl: String,
   itemsPerPageOptions: {
@@ -30,13 +30,6 @@ const items = ref([]);
 const pageNumber = ref(1);
 const itemsPerPage = ref(props.itemsPerPage);
 const totalCount = ref(0);
-const showPagination = computed(() => {
-  return props.showPagination && numberOfPages.value > 1;
-});
-
-const numberOfPages = computed(() => {
-  return Math.ceil(totalCount.value / itemsPerPage.value);
-});
 
 const showItemsPerPage = computed(() => {
   return props.showItemsPerPage;
@@ -44,42 +37,6 @@ const showItemsPerPage = computed(() => {
 
 function changeItemsPerPage(value) {
   itemsPerPage.value = value;
-  loadData();
-}
-
-function calculatePageNumbers() {
-  if (numberOfPages.value <= 10) {
-    return Array.from(Array(numberOfPages.value), (_, index) => index + 1);
-  }
-
-  let startIndex = pageNumber.value - 5;
-  let endIndex = pageNumber.value + 5;
-  if (startIndex < 0) {
-    // Increase the end index by the offset
-    endIndex -= startIndex;
-    startIndex = 0;
-  }
-
-  if (endIndex > numberOfPages.value) {
-    endIndex = numberOfPages.value;
-  }
-
-  return Array.from(Array(endIndex - startIndex), (_, index) => index + startIndex + 1);
-}
-
-function nextPage() {
-  pageNumber.value = pageNumber.value + 1;
-  if (pageNumber.value > numberOfPages.value) {
-    pageNumber.value = numberOfPages.value;
-  }
-  loadData();
-}
-
-function previousPage() {
-  pageNumber.value = pageNumber.value - 1;
-  if (pageNumber.value == 0) {
-    pageNumber.value = 1;
-  }
   loadData();
 }
 
@@ -142,31 +99,7 @@ onUnmounted(() => {
         </ul>
       </div>
     </div>
-    <div v-if="showPagination" class="col align-self-center">
-      <ul class="pagination justify-content-center">
-        <li class="page-item" :class="{ disabled: pageNumber === 1 }">
-          <a class="page-link" href="#" @click.prevent="previousPage">Previous</a>
-        </li>
-        <li v-if="pageNumber > 5 && numberOfPages > 10" class="page-item">
-          <a @click.prevent="setPage(1)" class="page-link" href="#">1</a>
-        </li>
-        <li v-if="pageNumber > 5 && numberOfPages > 10" class="page-item">
-          <a @click.prevent="setPage(pageNumber - 5)" class="page-link" href="#">...</a>
-        </li>
-        <li v-for="n in calculatePageNumbers()" class="page-item" :class="{ active: pageNumber === n }" :key="n">
-          <a @click.prevent="setPage(n)" class="page-link" href="#">{{ n }}</a>
-        </li>
-        <li v-if="numberOfPages - pageNumber > 5" class="page-item">
-          <a @click.prevent="setPage(pageNumber + 5)" class="page-link" href="#">...</a>
-        </li>
-        <li v-if="numberOfPages - pageNumber > 5" class="page-item">
-          <a @click.prevent="setPage(numberOfPages)" class="page-link" href="#">{{ numberOfPages }}</a>
-        </li>
-        <li class="page-item" :class="{ disabled: pageNumber === numberOfPages }">
-          <a class="page-link" href="#" @click.prevent="nextPage">Next</a>
-        </li>
-      </ul>
-    </div>
+    <Pagination :totalCount="totalCount" :itemsPerPage="itemsPerPage" :pageNumber="pageNumber" :showPagination="showPagination" @pageChanged="setPage" />
   </div>
   <slot name="footer" :count="totalCount"></slot>
 </template>
